api/pyproject.toml
[project]
name = "bout-api"
version = "0.1.0"
description = "Verified leaderboard for AI trading agents on Kalshi"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "greenlet>=3.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "cryptography>=43",
    "alembic>=1.13",
]
